Fix two lgtm.com alerts: comparisons using '==' and '!=' operators only test for reference equality
More details:
 - AbstractDirectoryEntry: https://lgtm.com/projects/g/elki-project/elki/snapshot/6012705b63d63424bed08d3f4c459c6149d64115/files/elki-index/src/main/java/de/lmu/ifi/dbs/elki/index/tree/AbstractDirectoryEntry.java#L131
 - CASHInterval: https://lgtm.com/projects/g/elki-project/elki/snapshot/6012705b63d63424bed08d3f4c459c6149d64115/files/elki-index/src/main/java/de/lmu/ifi/dbs/elki/index/tree/AbstractDirectoryEntry.java#L131

A total of 160 alerts have been flagged up by lgtm.com: https://lgtm.com/projects/g/elki-project/elki/alerts/

Tip: enable pull request integration for fully automated PR reviews to
make sure tiny mistakes like these do not actually get past code review!

+++ elki-index/src/main/java/de/lmu/ifi/dbs/elki/index/tree/AbstractDirectoryEntry.java
@@ -128,7 +128,7 @@ public abstract class AbstractDirectoryEntry implements DirectoryEntry {
 
     final AbstractDirectoryEntry that = (AbstractDirectoryEntry) o;
 
-    return id == that.id;
+    return id.intValue() == that.id.intValue();
   }
 
   /**
